refactor the job modification methods
