+++ plugins/nf-nomad/src/main/nextflow/nomad/executor/NomadService.groovy
@@ -72,20 +72,6 @@ class NomadService implements Closeable{
         this.variablesApi = new VariablesApi(apiClient)
     }
 
-    protected Resources getResources(TaskRun task) {
-        final DEFAULT_CPUS = 1
-        final DEFAULT_MEMORY = "500.MB"
-
-        final taskCfg = task.getConfig()
-        final taskCores =  !taskCfg.get("cpus") ? DEFAULT_CPUS :  taskCfg.get("cpus") as Integer
-        final taskMemory = taskCfg.get("memory") ? new MemoryUnit( taskCfg.get("memory") as String ) : new MemoryUnit(DEFAULT_MEMORY)
-
-        final res = new Resources()
-                .cores(taskCores)
-                .memoryMB(taskMemory.toMega() as Integer)
-
-        return res
-    }
 
     @Override
     void close() throws IOException {
@@ -98,13 +84,13 @@ class NomadService implements Closeable{
                 .id(id)
                 .name(task.name)
                 .type("batch")
-                .datacenters(this.config.jobOpts().datacenters)
+//                .datacenters(task, this.config.jobOpts().datacenters)
                 .namespace(this.config.jobOpts().namespace)
-                .taskGroups([createTaskGroup(task, args, env)])
+                .taskGroups([JobBuilder.createTaskGroup(task, args, env, this.config.jobOpts())])
                 .build()
 
-        assignDatacenters(task, job)
-        spreads(task, job)
+        JobBuilder.assignDatacenters(task, job)
+        JobBuilder.spreads(task, job, this.config.jobOpts())
 
         JobRegisterRequest jobRegisterRequest = new JobRegisterRequest()
         jobRegisterRequest.setJob(job)
@@ -127,215 +113,6 @@ class NomadService implements Closeable{
         }
     }
 
-    TaskGroup createTaskGroup(TaskRun taskRun, List<String> args, Map<String, String>env){
-        final ReschedulePolicy taskReschedulePolicy  = new ReschedulePolicy().attempts(this.config.jobOpts().rescheduleAttempts)
-        final RestartPolicy taskRestartPolicy  = new RestartPolicy().attempts(this.config.jobOpts().restartAttempts)
-
-        def task = createTask(taskRun, args, env)
-        def taskGroup = new TaskGroup(
-                name: "group",
-                tasks: [ task ],
-                reschedulePolicy: taskReschedulePolicy,
-                restartPolicy: taskRestartPolicy
-        )
-
-
-        if( config.jobOpts().volumeSpec ) {
-            taskGroup.volumes = [:]
-            config.jobOpts().volumeSpec.eachWithIndex { volumeSpec , idx->
-                if (volumeSpec && volumeSpec.type == JobVolume.VOLUME_CSI_TYPE) {
-                    taskGroup.volumes["vol_${idx}".toString()] = new VolumeRequest(
-                            type: volumeSpec.type,
-                            source: volumeSpec.name,
-                            attachmentMode: volumeSpec.attachmentMode,
-                            accessMode: volumeSpec.accessMode,
-                            readOnly: volumeSpec.readOnly,
-                    )
-                }
-
-                if (volumeSpec && volumeSpec.type == JobVolume.VOLUME_HOST_TYPE) {
-                    taskGroup.volumes["vol_${idx}".toString()] = new VolumeRequest(
-                            type: volumeSpec.type,
-                            source: volumeSpec.name,
-                            readOnly: volumeSpec.readOnly,
-                    )
-                }
-            }
-        }
-        return taskGroup
-    }
-
-    Task createTask(TaskRun task, List<String> args, Map<String, String>env) {
-        final DRIVER = "docker"
-        final DRIVER_PRIVILEGED = true
-
-        final imageName = task.container
-        final workingDir = task.workDir.toAbsolutePath().toString()
-        final taskResources = getResources(task)
-
-
-        def taskDef = new Task(
-                name: "nf-task",
-                driver: DRIVER,
-                resources: taskResources,
-                config: [
-                        image     : imageName,
-                        privileged: DRIVER_PRIVILEGED,
-                        work_dir  : workingDir,
-                        command   : args.first(),
-                        args      : args.tail(),
-                ] as Map<String, Object>,
-                env: env,
-        )
-
-        volumes(task, taskDef, workingDir)
-        affinity(task, taskDef)
-        constraint(task, taskDef)
-        constraints(task, taskDef)
-        secrets(task, taskDef)
-        return taskDef
-    }
-
-    protected Task volumes(TaskRun task, Task taskDef, String workingDir){
-        if( config.jobOpts().dockerVolume){
-            String destinationDir = workingDir.split(File.separator).dropRight(2).join(File.separator)
-            taskDef.config.mount = [
-                    type : "volume",
-                    target : destinationDir,
-                    source : config.jobOpts().dockerVolume,
-                    readonly : false
-            ]
-        }
-
-        if( config.jobOpts().volumeSpec){
-            taskDef.volumeMounts = []
-            config.jobOpts().volumeSpec.eachWithIndex { volumeSpec, idx ->
-                String destinationDir = volumeSpec.workDir ?
-                        workingDir.split(File.separator).dropRight(2).join(File.separator) : volumeSpec.path
-                taskDef.volumeMounts.add new VolumeMount(
-                        destination: destinationDir,
-                        volume: "vol_${idx}".toString()
-                )
-            }
-        }
-
-        taskDef
-    }
-
-    protected Task affinity(TaskRun task, Task taskDef) {
-        if (config.jobOpts().affinitySpec) {
-            def affinity = new Affinity()
-            if (config.jobOpts().affinitySpec.attribute) {
-                affinity.ltarget(config.jobOpts().affinitySpec.attribute)
-            }
-
-            affinity.operand(config.jobOpts().affinitySpec.operator ?: "=")
-
-            if (config.jobOpts().affinitySpec.value) {
-                affinity.rtarget(config.jobOpts().affinitySpec.value)
-            }
-            if (config.jobOpts().affinitySpec.weight != null) {
-                affinity.weight(config.jobOpts().affinitySpec.weight)
-            }
-            taskDef.affinities([affinity])
-        }
-        taskDef
-    }
-
-    protected Task constraint(TaskRun task, Task taskDef){
-        if( config.jobOpts().constraintSpec ){
-            def constraint = new Constraint()
-            if(config.jobOpts().constraintSpec.attribute){
-                constraint.ltarget(config.jobOpts().constraintSpec.attribute)
-            }
-
-            constraint.operand(config.jobOpts().constraintSpec.operator ?: "=")
-
-            if(config.jobOpts().constraintSpec.value){
-                constraint.rtarget(config.jobOpts().constraintSpec.value)
-            }
-            taskDef.constraints([constraint])
-        }
-
-        taskDef
-    }
-
-    protected Task constraints(TaskRun task, Task taskDef){
-        def constraints = [] as List<Constraint>
-
-        if( config.jobOpts().constraintsSpec ){
-            def list = ConstraintsBuilder.constraintsSpecToList(config.jobOpts().constraintsSpec)
-            constraints.addAll(list)
-        }
-
-        if( task.processor?.config?.get(TaskDirectives.CONSTRAINTS) &&
-                task.processor?.config?.get(TaskDirectives.CONSTRAINTS) instanceof Closure) {
-            Closure closure = task.processor?.config?.get(TaskDirectives.CONSTRAINTS) as Closure
-            JobConstraints constraintsSpec = JobConstraints.parse(closure)
-            def list = ConstraintsBuilder.constraintsSpecToList(constraintsSpec)
-            constraints.addAll(list)
-        }
-
-        if( constraints.size()) {
-            taskDef.constraints(constraints)
-        }
-        taskDef
-    }
-
-    protected Task secrets(TaskRun task, Task taskDef){
-        if( config.jobOpts()?.secretOpts?.enabled) {
-            def secrets = task.processor?.config?.get(TaskDirectives.SECRETS)
-            if (secrets) {
-                Template template = new Template(envvars: true, destPath: "/secrets/nf-nomad")
-                String secretPath = config.jobOpts()?.secretOpts?.path
-                String tmpl = secrets.collect { String name ->
-                    "${name}={{ with nomadVar \"$secretPath/${name}\" }}{{ .${name} }}{{ end }}"
-                }.join('\n').stripIndent()
-                template.embeddedTmpl(tmpl)
-                taskDef.addTemplatesItem(template)
-            }
-        }
-        taskDef
-    }
-
-    protected Job assignDatacenters(TaskRun task, Job job){
-        def datacenters = task.processor?.config?.get(TaskDirectives.DATACENTERS)
-        if( datacenters ){
-            if( datacenters instanceof List<String>) {
-                job.datacenters( datacenters as List<String>)
-                return job;
-            }
-            if( datacenters instanceof Closure) {
-                String str = datacenters.call().toString()
-                job.datacenters( [str])
-                return job;
-            }
-            job.datacenters( [datacenters.toString()] as List<String>)
-            return job
-        }
-        job
-    }
-
-    protected Job spreads(TaskRun task, Job jobDef){
-        def spreads = [] as List<Spread>
-        if( config.jobOpts().spreadsSpec ){
-            def list = SpreadsBuilder.spreadsSpecToList(config.jobOpts().spreadsSpec)
-            spreads.addAll(list)
-        }
-        if( task.processor?.config?.get(TaskDirectives.SPREAD) &&
-                task.processor?.config?.get(TaskDirectives.SPREAD) instanceof Map) {
-            Map map = task.processor?.config?.get(TaskDirectives.SPREAD) as Map
-            JobSpreads spreadSpec = new JobSpreads()
-            spreadSpec.spread(map)
-            def list = SpreadsBuilder.spreadsSpecToList(spreadSpec)
-            spreads.addAll(list)
-        }
-
-        spreads.each{
-            jobDef.addSpreadsItem(it)
-        }
-        jobDef
-    }
 
     String getJobState(String jobId){
         try {
